refactor: dto 파일들 정리
