+++ backend/api/src/main/java/com/yat2/episode/mindmap/dto/request/MindmapNameUpdateReq.java
@@ -1,0 +1,8 @@
+package com.yat2.episode.mindmap.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+
+public record MindmapNameUpdateReq(
+        @NotBlank(message = "이름은 공백일 수 없습니다.") @Size(max = 43, message = "이름은 43자 이내여야 합니다.") String name
+) {}
