+++ backend/api/src/main/java/com/yat2/episode/mindmap/dto/response/MindmapSessionJoinRes.java
@@ -1,4 +1,4 @@
-package com.yat2.episode.mindmap.dto;
+package com.yat2.episode.mindmap.dto.response;
 
 public record MindmapSessionJoinRes(
         String token,
