+++ backend/api/src/test/java/com/yat2/episode/mindmap/MindmapServiceTest.java
@@ -20,9 +20,9 @@ import com.yat2.episode.episode.EpisodeStarRepository;
 import com.yat2.episode.global.exception.CustomException;
 import com.yat2.episode.global.exception.ErrorCode;
 import com.yat2.episode.mindmap.constants.MindmapConstants;
-import com.yat2.episode.mindmap.dto.MindmapCreateReq;
-import com.yat2.episode.mindmap.dto.MindmapSessionJoinRes;
-import com.yat2.episode.mindmap.dto.MindmapSummaryRes;
+import com.yat2.episode.mindmap.dto.request.MindmapCreateReq;
+import com.yat2.episode.mindmap.dto.response.MindmapSessionJoinRes;
+import com.yat2.episode.mindmap.dto.response.MindmapSummaryRes;
 import com.yat2.episode.mindmap.jwt.MindmapJwtProperties;
 import com.yat2.episode.mindmap.jwt.MindmapJwtProvider;
 import com.yat2.episode.mindmap.jwt.MindmapTicketPayload;
